+++ modules/Dating_LSD2.py
@@ -29,15 +29,18 @@ class Dating_LSD2(Dating):
         if not isfile(sample_times_filename):
             raise ValueError("Invalid sample times file: %s" % sample_times_filename)
         lsd2_dir = '%s/LSD2' % GC.OUT_DIR_TMPFILES
-        makedirs(lsd2_dir, exist_ok=True)
         out_filename = '%s/dated.tre' % GC.OUT_DIR_OUTFILES
-        log_filename = '%s/log.txt' % lsd2_dir
-        lsd_times_filename = '%s/times_lsd.txt' % lsd2_dir
-        f = open(lsd_times_filename, 'w'); f.write(GC.convert_dates_LSD(sample_times_filename)); f.close()
-        command = ['lsd2', '-c', '-i', rooted_tree_filename, '-d', lsd_times_filename]
-        f = open('%s/command.txt' % lsd2_dir, 'w'); f.write('%s\n' % ' '.join(command)); f.close()
-        o = check_output(command)
-        move('%s.result' % rooted_tree_filename, out_filename)
-        for f in glob('%s.*' % rooted_tree_filename):
-            move(f, '%s/%s' % (lsd2_dir, f.split('/')[-1]))
+        if isfile(out_filename):
+            GC.SELECTED['Logging'].writeln("Dated tree exists. Skipping recomputation.")
+        else:
+            makedirs(lsd2_dir, exist_ok=True)
+            log_filename = '%s/log.txt' % lsd2_dir
+            lsd_times_filename = '%s/times_lsd.txt' % lsd2_dir
+            f = open(lsd_times_filename, 'w'); f.write(GC.convert_dates_LSD(sample_times_filename)); f.close()
+            command = ['lsd2', '-c', '-i', rooted_tree_filename, '-d', lsd_times_filename]
+            f = open('%s/command.txt' % lsd2_dir, 'w'); f.write('%s\n' % ' '.join(command)); f.close()
+            o = check_output(command)
+            move('%s.result' % rooted_tree_filename, out_filename)
+            for f in glob('%s.*' % rooted_tree_filename):
+                move(f, '%s/%s' % (lsd2_dir, f.split('/')[-1]))
         return out_filename
