+++ modules/MultipleSequenceAlignment_FSA.py
@@ -21,17 +21,19 @@ class MultipleSequenceAlignment_FSA(MultipleSequenceAlignment):
     def blurb():
         return "Multiple sequence alignment was performed using FSA (Bradley et al., 2009)."
 
-    def align(seqs_filename):
+    def align(seqs_filename, ref_id):
         if not isfile(seqs_filename):
             raise ValueError("Invalid sequence file: %s" % seqs_filename)
         fsa_dir = '%s/FSA' % GC.OUT_DIR_TMPFILES
-        makedirs(fsa_dir, exist_ok=True)
-        #f_stderr = open('%s/log.txt' % fsa_dir, 'w')
         out_filename = '%s/%s.aln' % (GC.OUT_DIR_OUTFILES, '.'.join(seqs_filename.split('/')[-1].split('.')[:-1]))
-        log_filename = '%s/log.txt' % fsa_dir
-        err = open('%s/err.txt' % fsa_dir, 'w')
-        out = open(out_filename, 'w')
-        command = ['fsa', '--logfile', log_filename, seqs_filename]
-        f = open('%s/command.txt' % fsa_dir, 'w'); f.write('%s\n' % ' '.join(command)); f.close()
-        call(command, stdout=out, stderr=err); out.close(); err.close()
+        if isfile(out_filename):
+            GC.SELECTED['Logging'].writeln("Multiple sequence alignment exists. Skipping recomputation.")
+        else:
+            makedirs(fsa_dir, exist_ok=True)
+            log_filename = '%s/log.txt' % fsa_dir
+            err = open('%s/err.txt' % fsa_dir, 'w')
+            out = open(out_filename, 'w')
+            command = ['fsa', '--logfile', log_filename, seqs_filename]
+            f = open('%s/command.txt' % fsa_dir, 'w'); f.write('%s\n' % ' '.join(command)); f.close()
+            call(command, stdout=out, stderr=err); out.close(); err.close()
         return out_filename
