+++ modules/MultipleSequenceAlignment_MUSCLE.py
@@ -21,14 +21,17 @@ class MultipleSequenceAlignment_MUSCLE(MultipleSequenceAlignment):
     def blurb():
         return "Multiple sequence alignment was performed using MUSCLE (Edgar, 2004)."
 
-    def align(seqs_filename):
+    def align(seqs_filename, ref_id):
         if not isfile(seqs_filename):
             raise ValueError("Invalid sequence file: %s" % seqs_filename)
         muscle_dir = '%s/MUSCLE' % GC.OUT_DIR_TMPFILES
-        makedirs(muscle_dir, exist_ok=True)
-        log_filename = '%s/log.txt' % muscle_dir
         out_filename = '%s/%s.aln' % (GC.OUT_DIR_OUTFILES, '.'.join(seqs_filename.split('/')[-1].split('.')[:-1]))
-        command = ['muscle', '-quiet', '-in', seqs_filename, '-out', out_filename, '-log', log_filename]
-        f = open('%s/command.txt' % muscle_dir, 'w'); f.write('%s\n' % ' '.join(command)); f.close()
-        call(command)
+        if isfile(out_filename):
+            GC.SELECTED['Logging'].writeln("Multiple sequence alignment exists. Skipping recomputation.")
+        else:
+            makedirs(muscle_dir, exist_ok=True)
+            log_filename = '%s/log.txt' % muscle_dir
+            command = ['muscle', '-quiet', '-in', seqs_filename, '-out', out_filename, '-log', log_filename]
+            f = open('%s/command.txt' % muscle_dir, 'w'); f.write('%s\n' % ' '.join(command)); f.close()
+            call(command)
         return out_filename
