+++ modules/MultipleSequenceAlignment_PRANK.py
@@ -22,15 +22,18 @@ class MultipleSequenceAlignment_PRANK(MultipleSequenceAlignment):
     def blurb():
         return "Multiple sequence alignment was performed using PRANK (Loytynoja, 2013)."
 
-    def align(seqs_filename):
+    def align(seqs_filename, ref_id):
         if not isfile(seqs_filename):
             raise ValueError("Invalid sequence file: %s" % seqs_filename)
         prank_dir = '%s/PRANK' % GC.OUT_DIR_TMPFILES
-        makedirs(prank_dir, exist_ok=True)
         out_filename = '%s/%s.aln' % (GC.OUT_DIR_OUTFILES, '.'.join(seqs_filename.split('/')[-1].split('.')[:-1]))
-        log = open('%s/log.txt' % prank_dir, 'w')
-        command = ['prank', '-o=%s' % out_filename, '-d=%s' % seqs_filename]
-        f = open('%s/command.txt' % prank_dir, 'w'); f.write('%s\n' % ' '.join(command)); f.close()
-        call(command, stdout=log); log.close()
-        move('%s.best.fas' % out_filename, out_filename)
+        if isfile(out_filename):
+            GC.SELECTED['Logging'].writeln("Multiple sequence alignment exists. Skipping recomputation.")
+        else:
+            makedirs(prank_dir, exist_ok=True)
+            log = open('%s/log.txt' % prank_dir, 'w')
+            command = ['prank', '-o=%s' % out_filename, '-d=%s' % seqs_filename]
+            f = open('%s/command.txt' % prank_dir, 'w'); f.write('%s\n' % ' '.join(command)); f.close()
+            call(command, stdout=log); log.close()
+            move('%s.best.fas' % out_filename, out_filename)
         return out_filename
