+++ modules/Preprocessing_SafeNames.py
@@ -41,45 +41,57 @@ class Preprocessing_SafeNames(Preprocessing):
             out_categories_filename = '%s/categories_safe.txt' % GC.OUT_DIR_OUTFILES
 
         # output safe sequences
-        f = open(out_seqs_filename, 'w')
-        for line in open(seqs_filename):
-            l = line.strip()
-            if len(l) == 0:
-                continue
-            elif l[0] == '>':
-                f.write('>'); f.write(GC.safe(l[1:]))
-            else:
-                f.write(l)
-            f.write('\n')
-        f.close()
-
-        # output safe sample times
-        f = open(out_times_filename, 'w')
-        for line in open(sample_times_filename):
-            parts = [v.strip() for v in line.strip().split('\t')]
-            if len(parts) == 2:
-                f.write(GC.safe(parts[0])); f.write('\t'); f.write(parts[1]); f.write('\n')
-            else:
-                raise ValueError("Invalid sample times file: %s" % sample_times_filename)
-        f.close()
-
-        # output safe outgroup names
-        if outgroups_filename is not None:
-            f = open(out_outgroups_filename, 'w')
-            for line in open(outgroups_filename):
-                f.write(GC.safe(line.strip())); f.write('\n')
+        if isfile(out_seqs_filename):
+            GC.SELECTED['Logging'].writeln("Safename sequences exist. Skipping recomputation.")
+        else:
+            f = open(out_seqs_filename, 'w')
+            for line in open(seqs_filename):
+                l = line.strip()
+                if len(l) == 0:
+                    continue
+                elif l[0] == '>':
+                    f.write('>'); f.write(GC.safe(l[1:]))
+                else:
+                    f.write(l)
+                f.write('\n')
             f.close()
 
-        # output safe categories
-        if categories_filename is not None:
-            f = open(out_categories_filename, 'w')
-            for line in open(categories_filename):
+        # output safe sample times
+        if isfile(out_times_filename):
+            GC.SELECTED['Logging'].writeln("Safename sample times exist. Skipping recomputation.")
+        else:
+            f = open(out_times_filename, 'w')
+            for line in open(sample_times_filename):
                 parts = [v.strip() for v in line.strip().split('\t')]
                 if len(parts) == 2:
                     f.write(GC.safe(parts[0])); f.write('\t'); f.write(parts[1]); f.write('\n')
                 else:
-                    raise ValueError("Invalid sample categories file: %s" % categories_filename)
+                    raise ValueError("Invalid sample times file: %s" % sample_times_filename)
             f.close()
+
+        # output safe outgroup names
+        if outgroups_filename is not None:
+            if isfile(out_outgroups_filename):
+                GC.SELECTED['Logging'].writeln("Safename outgroups exist. Skipping recomputation.")
+            else:
+                f = open(out_outgroups_filename, 'w')
+                for line in open(outgroups_filename):
+                    f.write(GC.safe(line.strip())); f.write('\n')
+                f.close()
+
+        # output safe categories
+        if categories_filename is not None:
+            if isfile(out_categories_filename):
+                GC.SELECTED['Logging'].writeln("Safename sample categories exist. Skipping recomputation.")
+            else:
+                f = open(out_categories_filename, 'w')
+                for line in open(categories_filename):
+                    parts = [v.strip() for v in line.strip().split('\t')]
+                    if len(parts) == 2:
+                        f.write(GC.safe(parts[0])); f.write('\t'); f.write(parts[1]); f.write('\n')
+                    else:
+                        raise ValueError("Invalid sample categories file: %s" % categories_filename)
+                f.close()
 
         # return output filenames
         return out_seqs_filename, out_times_filename, out_outgroups_filename, out_categories_filename
